Attempt to use json.load(fdesc) again
This PR includes:
- Use json.load(fdesc)
- Early exit in tokenresolver
- Move caching functionality to utils.py
- Fix 2 issues
- Return default value on failure
- Fix more json.load() calls for Python 3.5
