+++ resources/lib/resumepoints.py
@@ -12,9 +12,9 @@ except ImportError:  # Python 2
     from urllib2 import build_opener, install_opener, ProxyHandler, Request, HTTPError, urlopen
 
 from data import SECONDS_MARGIN
-from kodiutils import (container_refresh, get_proxies, get_setting, has_credentials, input_down,
-                       localize, log, log_error, notification)
-from utils import get_cache, get_url_json, invalidate_caches, update_cache
+from kodiutils import (container_refresh, get_cache, get_proxies, get_setting, get_url_json,
+                       has_credentials, input_down, invalidate_caches, localize, log, log_error,
+                       notification, update_cache)
 
 
 class ResumePoints:
@@ -46,7 +46,7 @@ class ResumePoints:
                 }
                 resumepoints_url = 'https://video-user-data.vrt.be/resume_points'
                 resumepoints_json = get_url_json(url=resumepoints_url, cache='resume_points.json', headers=headers)
-        if resumepoints_json:
+        if resumepoints_json is not None:
             self._resumepoints = resumepoints_json
 
     def update(self, asset_id, title, url, watch_later=None, position=None, total=None, whatson_id=None, asynchronous=False):
@@ -69,7 +69,7 @@ class ResumePoints:
             # resumepoint is not changed, nothing to do
             return True
 
-        from statichelper import reformat_url
+        from utils import reformat_url
         url = reformat_url(url, 'short')
 
         if asset_id in self._resumepoints:
@@ -171,7 +171,7 @@ class ResumePoints:
 
     def get_url(self, asset_id, url_type='medium'):
         ''' Return the stored url a video '''
-        from statichelper import reformat_url
+        from utils import reformat_url
         return reformat_url(self._resumepoints.get(asset_id, {}).get('value', {}).get('url'), url_type)
 
     @staticmethod
