+++ resources/lib/utils.py
@@ -1,183 +1,210 @@
 # -*- coding: utf-8 -*-
 # GNU General Public License v3.0 (see COPYING or https://www.gnu.org/licenses/gpl-3.0.txt)
-''' All functionality that requires Kodi imports '''
+''' Implements static functions used elsewhere in the add-on '''
 
 from __future__ import absolute_import, division, unicode_literals
-from sys import version_info
+import re
 
 try:  # Python 3
-    from urllib.error import HTTPError
-    from urllib.parse import unquote
-    from urllib.request import urlopen, Request
+    from html import unescape
 except ImportError:  # Python 2
-    from urllib2 import HTTPError, unquote, urlopen, Request
+    from HTMLParser import HTMLParser
 
-from kodiutils import (container_refresh, delete, exists, get_cache_path, get_setting, listdir,
-                       localize, log, log_error, mkdirs, notification, ok_dialog, open_file, stat_file)
-from statichelper import to_unicode
+    def unescape(string):
+        ''' Expose HTMLParser's unescape '''
+        return HTMLParser().unescape(string)
 
-
-def human_delta(seconds):
-    ''' Return a human-readable representation of the TTL '''
-    from math import floor
-    days = int(floor(seconds / (24 * 60 * 60)))
-    seconds = seconds % (24 * 60 * 60)
-    hours = int(floor(seconds / (60 * 60)))
-    seconds = seconds % (60 * 60)
-    if days:
-        return '%d day%s and %d hour%s' % (days, 's' if days != 1 else '', hours, 's' if hours != 1 else '')
-    minutes = int(floor(seconds / 60))
-    seconds = seconds % 60
-    if hours:
-        return '%d hour%s and %d minute%s' % (hours, 's' if hours != 1 else '', minutes, 's' if minutes != 1 else '')
-    if minutes:
-        return '%d minute%s and %d second%s' % (minutes, 's' if minutes != 1 else '', seconds, 's' if seconds != 1 else '')
-    return '%d second%s' % (seconds, 's' if seconds != 1 else '')
+HTML_MAPPING = [
+    (re.compile(r'<(/?)i(|\s[^>]+)>', re.I), '[\\1I]'),
+    (re.compile(r'<(/?)b(|\s[^>]+)>', re.I), '[\\1B]'),
+    (re.compile(r'<em(|\s[^>]+)>', re.I), '[B][COLOR yellow]'),
+    (re.compile(r'</em>', re.I), '[/COLOR][/B]'),
+    (re.compile(r'<li>', re.I), '- '),
+    (re.compile(r'</?(div|li|p|span|ul)(|\s[^>]+)>', re.I), ''),
+    (re.compile('<br>\n{0,1}', re.I), ' '),  # This appears to be specific formatting for VRT NU, but unwanted by us
+    (re.compile('(&nbsp;\n){2,}', re.I), '\n'),  # Remove repeating non-blocking spaced newlines
+]
 
 
-def get_cache(path, ttl=None):  # pylint: disable=redefined-outer-name
-    ''' Get the content from cache, if it's still fresh '''
-    if get_setting('usehttpcaching', 'true') == 'false':
+def to_unicode(text, encoding='utf-8', errors='strict'):
+    ''' Force text to unicode '''
+    if isinstance(text, bytes):
+        return text.decode(encoding, errors=errors)
+    return text
+
+
+def from_unicode(text, encoding='utf-8', errors='strict'):
+    ''' Force unicode to text '''
+    import sys
+    if sys.version_info.major == 2 and isinstance(text, unicode):  # noqa: F821; pylint: disable=undefined-variable
+        return text.encode(encoding, errors)
+    return text
+
+
+def capitalize(string):
+    ''' Ensure the first character is uppercase '''
+    string = string.strip()
+    return string[0].upper() + string[1:]
+
+
+def strip_newlines(text):
+    ''' Strip newlines and whitespaces '''
+    return text.replace('\n', '').strip()
+
+
+def html_to_kodilabel(text):
+    ''' Convert VRT HTML content into Kodit formatted text '''
+    for key, val in HTML_MAPPING:
+        text = key.sub(val, text)
+    return unescape(text).strip()
+
+
+def reformat_url(url, url_type):
+    ''' Convert a url '''
+    # Clean URLs with a hash in it
+    pos = url.find('#')
+    if pos >= 0:
+        url = url[:pos]
+    # long url
+    if url_type == 'long':
+        if url.startswith('/vrtnu/a-z'):
+            return 'https://www.vrt.be' + url
+        if url.startswith('//www.vrt.be'):
+            return 'https:' + url
+        return url
+    # medium url
+    if url_type == 'medium':
+        if url.startswith('https:'):
+            return url.replace('https:', '')
+        if url.startswith('/vrtnu/a-z'):
+            return '//www.vrt.be' + url
+        return url
+    # short url
+    if url_type == 'short':
+        if url.startswith('https://www.vrt.be'):
+            return url.replace('https://www.vrt.be', '')
+        if url.startswith('//www.vrt.be'):
+            return url.replace('//www.vrt.be', '')
+    return url
+
+
+def program_to_url(program, url_type):
+    ''' Convert a program url component (e.g. de-campus-cup) to:
+        - a short programUrl (e.g. /vrtnu/a-z/de-campus-cup/)
+        - a medium programUrl (e.g. //www.vrt.be/vrtnu/a-z/de-campus-cup/)
+        - a long programUrl (e.g. https://www.vrt.be/vrtnu/a-z/de-campus-cup/)
+    '''
+    url = None
+    if program:
+        # short programUrl
+        if url_type == 'short':
+            url = '/vrtnu/a-z/' + program + '/'
+        # medium programUrl
+        elif url_type == 'medium':
+            url = '//www.vrt.be/vrtnu/a-z/' + program + '/'
+        # long programUrl
+        elif url_type == 'long':
+            url = 'https://www.vrt.be/vrtnu/a-z/' + program + '/'
+    return url
+
+
+def url_to_program(url):
+    ''' Convert
+          - a targetUrl (e.g. //www.vrt.be/vrtnu/a-z/de-campus-cup.relevant/),
+          - a short programUrl (e.g. /vrtnu/a-z/de-campus-cup/) or
+          - a medium programUrl (e.g. //www.vrt.be/vrtnu/a-z/de-campus-cup/)
+          - a long programUrl (e.g. https://www.vrt.be/vrtnu/a-z/de-campus-cup/)
+        to a program url component (e.g. de-campus-cup).
+        Any season or episode information is removed as well.
+    '''
+    program = None
+    if url.startswith('https://www.vrt.be/vrtnu/a-z/'):
+        # long programUrl or targetUrl
+        program = url.split('/')[5]
+    elif url.startswith('//www.vrt.be/vrtnu/a-z/'):
+        # medium programUrl or targetUrl
+        program = url.split('/')[5]
+    elif url.startswith('/vrtnu/a-z/'):
+        # short programUrl
+        program = url.split('/')[3]
+    if program.endswith('.relevant'):
+        # targetUrl
+        program = program.replace('.relevant', '')
+    return program
+
+
+def url_to_episode(url):
+    ''' Convert a targetUrl (e.g. //www.vrt.be/vrtnu/a-z/buck/1/buck-s1a32/) to
+        a short episode url (/vrtnu/a-z/buck/1/buck-s1a32/)
+    '''
+    if url.startswith('https://www.vrt.be/vrtnu/a-z/'):
+        # long episode url
+        return url.replace('https://www.vrt.be/vrtnu/a-z/', '/vrtnu/a-z/')
+    if url.startswith('//www.vrt.be/vrtnu/a-z/'):
+        # medium episode url
+        return url.replace('//www.vrt.be/vrtnu/a-z/', '/vrtnu/a-z/')
+    return None
+
+
+def video_to_api_url(url):
+    ''' Convert a full VRT NU url (e.g. https://www.vrt.be/vrtnu/a-z/de-ideale-wereld/2019-nj/de-ideale-wereld-d20191010/)
+        to a VRT Search API url (e.g. //www.vrt.be/vrtnu/a-z/de-ideale-wereld/2019-nj/de-ideale-wereld-d20191010/)
+    '''
+    if url.startswith('https:'):
+        url = url.replace('https:', '')
+        # NOTE: add a trailing slash again because routing plugin removes it and VRT NU Search API needs it
+        if not url.endswith('/'):
+            url += '/'
+    return url
+
+
+def play_url_to_id(url):
+    ''' Convert a plugin:// url (e.g. plugin://plugin.video.vrt.nu/play/id/vid-5b12c0f6-b8fe-426f-a600-557f501f3be9/pbs-pub-7e2764cf-a8c0-4e78-9cbc-46d39381c237)
+        to an id dictionary (e.g. {'video_id': 'vid-5b12c0f6-b8fe-426f-a600-557f501f3be9'}
+    '''
+    play_id = dict()
+    if 'play/id/' in url:
+        play_id['video_id'] = url.split('play/id/')[1].split('/')[0]
+    elif 'play/upnext/' in url:
+        play_id['video_id'] = url.split('play/upnext/')[1]
+    elif '/play/url/' in url:
+        play_id['video_url'] = video_to_api_url(url.split('play/url/')[1])
+    return play_id
+
+
+def shorten_link(url):
+    ''' Create a link that is as short as possible '''
+    if url is None:
         return None
-
-    fullpath = get_cache_path() + path
-    if not exists(fullpath):
-        return None
-
-    from time import localtime, mktime
-    mtime = stat_file(fullpath).st_mtime()
-    now = mktime(localtime())
-    if ttl and now >= mtime + ttl:
-        return None
-
-    if ttl is None:
-        log(3, "Cache '{path}' is forced from cache.", path=path)
-    else:
-        log(3, "Cache '{path}' is fresh, expires in {time}.", path=path, time=human_delta(mtime + ttl - now))
-    from json import load
-    with open_file(fullpath, 'r') as fdesc:
-        try:
-            return load(fdesc)
-        except (TypeError, ValueError) as exc:  # No JSON object could be decoded
-            fdesc.seek(0, 0)
-            log_error('{exc}\nDATA: {data}', exc=exc, data=fdesc.read())
-            return None
+    if url.startswith('https://www.vrt.be/vrtnu/'):
+        # As used in episode search result 'permalink'
+        return url.replace('https://www.vrt.be/vrtnu/', 'vrtnu.be/')
+    if url.startswith('//www.vrt.be/vrtnu/'):
+        # As used in program a-z listing 'targetUrl'
+        return url.replace('//www.vrt.be/vrtnu/', 'vrtnu.be/')
+    return url
 
 
-def update_cache(path, data):
-    ''' Update the cache, if necessary '''
-    if get_setting('usehttpcaching', 'true') == 'false':
-        return
-
-    from hashlib import md5
-    from json import dump, dumps
-    fullpath = get_cache_path() + path
-    if exists(fullpath):
-        with open_file(fullpath) as fdesc:
-            cachefile = fdesc.read().encode('utf-8')
-        md5_cache = md5(cachefile)
-    else:
-        md5_cache = 0
-        # Create cache directory if missing
-        if not exists(get_cache_path()):
-            mkdirs(get_cache_path())
-
-    # Avoid writes if possible (i.e. SD cards)
-    if md5_cache != md5(dumps(data).encode('utf-8')):
-        log(3, "Write cache '{path}'.", path=path)
-        with open_file(fullpath, 'w') as fdesc:
-            # dump(data, fdesc, encoding='utf-8')
-            dump(data, fdesc)
-    else:
-        # Update timestamp
-        from os import utime
-        log(3, "Cache '{path}' has not changed, updating mtime only.", path=path)
-        utime(path)
+def add_https_proto(url):
+    ''' Add HTTPS protocol to URL that lacks it '''
+    if url.startswith('//'):
+        return 'https:' + url
+    if url.startswith('/'):
+        return 'https://www.vrt.be' + url
+    return url
 
 
-def ttl(kind='direct'):
-    ''' Return the HTTP cache ttl in seconds based on kind of relation '''
-    if kind == 'direct':
-        return int(get_setting('httpcachettldirect', 5)) * 60
-    if kind == 'indirect':
-        return int(get_setting('httpcachettlindirect', 60)) * 60
-    return 5 * 60
-
-def get_json_data(response):
-    ''' Return json object from HTTP response '''
-    from json import load, loads
-    if (3, 0, 0) <= version_info <= (3, 5, 9):  # the JSON object must be str, not 'bytes'
-        json_data = loads(to_unicode(response.read()))
-    else:
-        json_data = load(response)
-    return json_data
-
-def get_url_json(url, cache=None, headers=None, data=None):
-    ''' Return HTTP data '''
-    if headers is None:
-        headers = dict()
-    log(2, 'URL get: {url}', url=unquote(url))
-    req = Request(url, headers=headers)
-    if data is not None:
-        req.data = data
+def realpage(page):
+    ''' Convert a URL parameter page value into an integer '''
     try:
-        json_data = get_json_data(urlopen(req))
-    except ValueError as exc:  # No JSON object could be decoded
-        log_error('JSON Error: {exc}', exc=exc)
-        return []
-    except HTTPError as exc:
-        if hasattr(req, 'selector'):  # Python 3.4+
-            url_length = len(req.selector)
-        else:  # Python 2.7
-            url_length = len(req.get_selector())
-        if exc.code == 413 and url_length > 8192:
-            ok_dialog(heading='HTTP Error 413', message=localize(30967))
-            log_error('HTTP Error 413: Exceeded maximum url length: '
-                      'VRT Search API url has a length of {length} characters.', length=url_length)
-            return []
-        if exc.code == 400 and 7600 <= url_length <= 8192:
-            ok_dialog(heading='HTTP Error 400', message=localize(30967))
-            log_error('HTTP Error 400: Probably exceeded maximum url length: '
-                      'VRT Search API url has a length of {length} characters.', length=url_length)
-            return []
-        try:
-            return get_json_data(exc)
-        except ValueError as exc:  # No JSON object could be decoded
-            log_error('JSON Error: {exc}', exc=exc)
-            return []
-        raise
-    else:
-        if cache:
-            update_cache(cache, json_data)
-    return json_data
+        page = int(page)
+    except ValueError:
+        return 1
+    if page < 1:
+        return 1
+    return page
 
 
-def get_cached_url_json(url, cache, headers=None, ttl=None):  # pylint: disable=redefined-outer-name
-    ''' Return data from cache, if any, else make an HTTP request '''
-    # Get api data from cache if it is fresh
-    json_data = get_cache(cache, ttl=ttl)
-    if json_data is not None:
-        return json_data
-    return get_url_json(url, cache=cache, headers=headers)
-
-
-def refresh_caches(cache_file=None):
-    ''' Invalidate the needed caches and refresh container '''
-    files = ['favorites.json', 'oneoff.json', 'resume_points.json']
-    if cache_file and cache_file not in files:
-        files.append(cache_file)
-    invalidate_caches(*files)
-    container_refresh()
-    notification(message=localize(30981))
-
-
-def invalidate_caches(*caches):
-    ''' Invalidate multiple cache files '''
-    import fnmatch
-    _, files = listdir(get_cache_path())
-    # Invalidate caches related to menu list refreshes
-    removes = set()
-    for expr in caches:
-        removes.update(fnmatch.filter(files, expr))
-    for filename in removes:
-        delete(get_cache_path() + filename)
+def find_entry(dlist, key, value, default=None):
+    ''' Find (the first) dictionary in a list where key matches value '''
+    return next((entry for entry in dlist if entry.get(key) == value), default)
